Fix bugs uncovered with the increase in test coverage
* Potential `nil` pointer dereference when the outbound HTTP request experienced
  an error (request failed to complete).
* Mistakenly expecting 403 to be the status code used when the API key provided
  is invalid.

The first issue was that the `*http.Response` returned from `*http.Client.Do()`
was being used in the error handling block for that function call to get the
request's URL. This was updated to properly use `req.URL`.

The second issue was simply catching the wrong HTTP status code on request
processing. This updates the code, and tests, to use http.StatusUnauthorized for
authentication failures.

Signed-off-by: Tim Heckman <t@heckman.io>