Remove DecodeIP, enhance error handling, remove anonymous usage
This removes the DecodeIP function from the package, updates the decoding of
request data to be safer, and changes the error propagation mechanism to more
closely match errors returned from the API.

The first is to remove the DecodeIP function that was just unmarshaling JSON. If
you were using this in your project, you can replace the function call with a
standard `ioutil.ReadAll()` + `json.Unmarshal()`.

Any API error is now returned as an ipdata.Error type value. This includes a
`Code()` method which returns the numeric HTTP status code, and `Error()` which
returns the error message sent by the API.

Lastly, this removes the support for anonymous API usage. This means an apiKey
is a required parameter for generating a client. As a result, NewClient now
returns an error value which was a breaking API change.