Add support for Bulk Lookup operations and prep for v0.7.0
This change adds the `RawBulkLookup()` and `BulkLookup()` methods to the ipdata
Client type. The functions take a slice of IPs in string form, and returns a
slice of IP structs.

If there is an error looking up any of the IPs in the slice, the first error
encountered is returned and the IP slice will contain data (as long as one
lookup in the bulk action succeeded). Any items that errored will be a nil value
within the slice, so be careful