Add context to all requests
This adds the possibility to provide a context.Context to all GET
and POST requests and introduces:

- LookupWithContext calls RawLookupWithContext calls newGetRequestWithContext
- BulkLookupWithContext calls RawBulkLookupWithContext calls newPostRequestWithContext

The new functions are duplicated from the old ones but use a NewRequestWithContext instead.
The old functions call the WithContext functions but provide a context.Background(), so
only one set needs to be maintained.